+++ plugins/test/railroad/plugininfo.d
@@ -1,0 +1,6 @@
+module test.railroad.plugininfo;
+enum id = "test.railroad";
+enum semver = "0.7.0";
+enum deps = [];
+enum clientdeps = [];
+enum serverdeps = [];